third_party/WebKit/LayoutTests/svg/animations/animateMotion-multiple-expected.txt

SVG 1.1 dynamic animation tests

Test for checking position of the svg element when multiple animateMotion are acting on it

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getBBox().x is 20
PASS rootSVGElement.getBBox().x is 20
PASS rootSVGElement.getBBox().x is 40
PASS rootSVGElement.getBBox().x is 60
PASS rootSVGElement.getBBox().x is 20
PASS rootSVGElement.getBBox().x is 20
PASS successfullyParsed is true

TEST COMPLETE
